feat(backend): auth JWT, flujos documentales, emails y upload de arch…

+++ src/omnifiles_bd_v4.sql
@@ -18,24 +18,24 @@ USE omnifiles;
 -- ============================================================
 
 CREATE TABLE rol (
-    id      INT AUTO_INCREMENT PRIMARY KEY,
+    id      BIGINT AUTO_INCREMENT PRIMARY KEY,
     nombre  VARCHAR(50) NOT NULL UNIQUE
 );
 
 CREATE TABLE estado_documento (
-    id          INT AUTO_INCREMENT PRIMARY KEY,
+    id          BIGINT AUTO_INCREMENT PRIMARY KEY,
     nombre      VARCHAR(50)  NOT NULL UNIQUE COMMENT 'Clave Java: CREADO | EN_REVISION | APROBADO | RECHAZADO',
     descripcion VARCHAR(200)
 );
 
 CREATE TABLE estado_tarea (
-    id          INT AUTO_INCREMENT PRIMARY KEY,
+    id          BIGINT AUTO_INCREMENT PRIMARY KEY,
     nombre      VARCHAR(50)  NOT NULL UNIQUE COMMENT 'Clave Java: PENDIENTE | APROBADO | RECHAZADO | CORRECCION',
     descripcion VARCHAR(200)
 );
 
 CREATE TABLE tipo_accion (
-    id          INT AUTO_INCREMENT PRIMARY KEY,
+    id          BIGINT AUTO_INCREMENT PRIMARY KEY,
     nombre      VARCHAR(50)  NOT NULL UNIQUE COMMENT 'Clave Java: CREACION | ACTUALIZACION | CAMBIO_ESTADO | ELIMINACION | RESTAURACION | DESCARGA',
     descripcion VARCHAR(200)
 );
@@ -45,19 +45,19 @@ CREATE TABLE tipo_accion (
 -- ============================================================
 
 CREATE TABLE usuario (
-    id          INT          AUTO_INCREMENT PRIMARY KEY,
+    id          BIGINT       AUTO_INCREMENT PRIMARY KEY,
     nombre      VARCHAR(100) NOT NULL,
     email       VARCHAR(100) NOT NULL UNIQUE,
     contrasena  VARCHAR(255) NOT NULL               COMMENT 'BCrypt — el backend hashea automáticamente',
     estado      BOOLEAN      NOT NULL DEFAULT TRUE  COMMENT 'TRUE = activo',
     telefono    VARCHAR(20)  NULL,
     cargo       VARCHAR(100) NULL,
-    rol_id      INT          NOT NULL,
+    rol_id      BIGINT          NOT NULL,
     CONSTRAINT fk_usuario_rol FOREIGN KEY (rol_id) REFERENCES rol(id)
 );
 
 CREATE TABLE tipo_documento (
-    id                INT          AUTO_INCREMENT PRIMARY KEY,
+    id                BIGINT       AUTO_INCREMENT PRIMARY KEY,
     nombre            VARCHAR(100) NOT NULL UNIQUE  COMMENT 'Nombre de la plantilla (ej: Recibo de pago, Contrato)',
     descripcion       TEXT,
     estado            BOOLEAN      NOT NULL DEFAULT TRUE  COMMENT 'TRUE = disponible para crear documentos',
@@ -67,8 +67,8 @@ CREATE TABLE tipo_documento (
 -- Flujo de aprobación asociado a una plantilla
 -- Una plantilla tiene exactamente un flujo
 CREATE TABLE flujo (
-    id                INT          AUTO_INCREMENT PRIMARY KEY,
-    tipo_documento_id INT          NOT NULL UNIQUE  COMMENT 'Cada plantilla tiene un solo flujo',
+    id                BIGINT       AUTO_INCREMENT PRIMARY KEY,
+    tipo_documento_id BIGINT          NOT NULL UNIQUE  COMMENT 'Cada plantilla tiene un solo flujo',
     nombre            VARCHAR(100) NOT NULL          COMMENT 'Nombre descriptivo del flujo',
     CONSTRAINT fk_flujo_tipo FOREIGN KEY (tipo_documento_id) REFERENCES tipo_documento(id)
 );
@@ -77,12 +77,12 @@ CREATE TABLE flujo (
 -- El campo usuario_id es opcional: si viene, solo ese usuario puede resolver la tarea;
 -- si es NULL, cualquier usuario con el rol indicado puede resolverla
 CREATE TABLE etapa_flujo (
-    id           INT          AUTO_INCREMENT PRIMARY KEY,
-    flujo_id     INT          NOT NULL,
+    id           BIGINT       AUTO_INCREMENT PRIMARY KEY,
+    flujo_id     BIGINT          NOT NULL,
     orden        INT          NOT NULL  COMMENT 'Secuencia de ejecución (1, 2, 3...)',
     nombre       VARCHAR(100) NOT NULL  COMMENT 'Nombre de la etapa (ej: Revisión, Firma, Aprobación final)',
-    rol_id       INT          NOT NULL  COMMENT 'Rol responsable de esta etapa',
-    usuario_id   INT          NULL      COMMENT 'Usuario específico (NULL = cualquiera con el rol)',
+    rol_id       BIGINT          NOT NULL  COMMENT 'Rol responsable de esta etapa',
+    usuario_id   BIGINT          NULL      COMMENT 'Usuario específico (NULL = cualquiera con el rol)',
     CONSTRAINT fk_etapa_flujo    FOREIGN KEY (flujo_id)   REFERENCES flujo(id),
     CONSTRAINT fk_etapa_rol      FOREIGN KEY (rol_id)     REFERENCES rol(id),
     CONSTRAINT fk_etapa_usuario  FOREIGN KEY (usuario_id) REFERENCES usuario(id),
@@ -90,15 +90,15 @@ CREATE TABLE etapa_flujo (
 );
 
 CREATE TABLE documento (
-    id                  INT          AUTO_INCREMENT PRIMARY KEY,
+    id                  BIGINT       AUTO_INCREMENT PRIMARY KEY,
     nombre              VARCHAR(150) NOT NULL,
     ruta_archivo        VARCHAR(500) NULL              COMMENT 'Ruta física del archivo subido',
-    estado_id           INT          NOT NULL,
+    estado_id           BIGINT          NOT NULL,
     fecha_creacion      DATETIME     NOT NULL,
     fecha_actualizacion DATETIME,
     eliminado           BOOLEAN      NOT NULL DEFAULT FALSE COMMENT 'TRUE = papelera',
-    usuario_id          INT          NOT NULL,
-    tipo_documento_id   INT,
+    usuario_id          BIGINT          NOT NULL,
+    tipo_documento_id   BIGINT,
     CONSTRAINT fk_documento_estado  FOREIGN KEY (estado_id)         REFERENCES estado_documento(id),
     CONSTRAINT fk_documento_usuario FOREIGN KEY (usuario_id)        REFERENCES usuario(id),
     CONSTRAINT fk_documento_tipo    FOREIGN KEY (tipo_documento_id) REFERENCES tipo_documento(id)
@@ -107,11 +107,11 @@ CREATE TABLE documento (
 -- Tarea: ahora incluye etapa_flujo_id para saber en qué paso del flujo está
 -- y poder avanzar automáticamente al siguiente cuando se aprueba
 CREATE TABLE tarea (
-    id                  INT          AUTO_INCREMENT PRIMARY KEY,
-    documento_id        INT          NOT NULL,
-    usuario_asignado_id INT          NOT NULL,
-    etapa_flujo_id      INT          NOT NULL  COMMENT 'Etapa del flujo a la que corresponde esta tarea',
-    estado_id           INT          NOT NULL,
+    id                  BIGINT       AUTO_INCREMENT PRIMARY KEY,
+    documento_id        BIGINT          NOT NULL,
+    usuario_asignado_id BIGINT          NOT NULL,
+    etapa_flujo_id      BIGINT          NOT NULL  COMMENT 'Etapa del flujo a la que corresponde esta tarea',
+    estado_id           BIGINT          NOT NULL,
     observaciones       VARCHAR(500),
     fecha_asignacion    DATETIME     NOT NULL,
     fecha_resolucion    DATETIME               COMMENT 'NULL mientras esté PENDIENTE',
@@ -122,13 +122,13 @@ CREATE TABLE tarea (
 );
 
 CREATE TABLE historial_documento (
-    id           INT      AUTO_INCREMENT PRIMARY KEY,
-    estado_id    INT      NOT NULL    COMMENT 'Estado en que quedó el documento',
-    accion_id    INT      NOT NULL    COMMENT 'Tipo de acción realizada',
+    id           BIGINT   AUTO_INCREMENT PRIMARY KEY,
+    estado_id    BIGINT      NOT NULL    COMMENT 'Estado en que quedó el documento',
+    accion_id    BIGINT      NOT NULL    COMMENT 'Tipo de acción realizada',
     descripcion  VARCHAR(300) NULL   COMMENT 'Detalle adicional (ej: etapa completada, quién aprobó)',
     fecha_cambio DATETIME NOT NULL,
-    documento_id INT      NOT NULL,
-    usuario_id   INT      NOT NULL,
+    documento_id BIGINT      NOT NULL,
+    usuario_id   BIGINT      NOT NULL,
     CONSTRAINT fk_historial_estado    FOREIGN KEY (estado_id)    REFERENCES estado_documento(id),
     CONSTRAINT fk_historial_accion    FOREIGN KEY (accion_id)    REFERENCES tipo_accion(id),
     CONSTRAINT fk_historial_documento FOREIGN KEY (documento_id) REFERENCES documento(id) ON DELETE CASCADE,
@@ -213,3 +213,4 @@ INSERT INTO etapa_flujo (flujo_id, orden, nombre, rol_id, usuario_id) VALUES
 
 UPDATE tipo_documento SET flujo_configurado = TRUE WHERE id = 2;
 */
+
